Remove erroneous GNSS receiver update
updateGNSSSettings was being forced to true no matter what. Removing. In the future, we should set updateGNSSSettings to true any time a given setting cannot be changed instantaneously and thus requires the GNSS receiver be sent through all its configuration functions at the next boot. The most common example of this is when the user exits Web Config: a large number of settings may have been changed, and therefor the GNSS should be configured after reset.