OTA: Support WiFi setup and firmware update as one change
Always display the update button and menu item, remove the requirement
check for new firmware qualification.  When the settings are written
back to to the GNSS device the system is able to establish the
connection to the remote access point using the newly specified WIFI
SSID and password and then perform the OTA firmware update.

Note: Due to the WiFi restart the connection is broken between the
browser and the GNSS device.  As a result, the browser does not receive
the status update message for the progress bar and reset.

Prior to this change the user would need to check for new firmware
before performing the firmware update:

1. Use a browser to connect to 192.168.4.1
2. Click on the "WiFi Configuration" button
3. Specify the WiFi SSID and Password
4. Click on the "Save Configuration" button
5. Reconnect the browser to 192.168.4.1
6. Click on the "System Configuration" button
7. Click on the "Check for New Firmware" button
8. If new firmware is available, click on the "Update to vXX.YY" button

While it is possible to change the WiFi SSID and password and then hit
the "Check for New Firmware" button the response gets lost because of
the WiFi restart.  When the browser is reconnected the "Update
to vXX.YY" button is still disabled because the page got reloaded and
the browser does not have the new firmware version.  As such, the "Check
for New Firmware" button needs to be pressed again to get the new
firmware version and enable the "Update to vXX.YY" button.