WiFi: Fix WiFi startup
Remove the chicken and egg issue where WiFi needs to be running to be
the highest priority, but has not started yet because it is not enabled!

The new routine, networkInterfaceRunning, allows any interface to poll
the network and determine if the interface should be running.  This
routine provides the missing feedback between the network and the
interface layers to when the network layer wants to stop the interface
and a stop sequence does not exist.